server/requirements.txt
click==8.1.3
colorama==0.4.6
dnspython==2.2.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-MQTT==1.1.1
flask-openapi3==2.2.0
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
paho-mqtt==1.6.1
pydantic==1.10.2
pymongo==4.3.3
python-dotenv==0.21.0
pytz==2022.6
six==1.16.0
typing_extensions==4.4.0
Werkzeug==2.2.2